Optimise chr bdlist handling

+++ src/include/game/chraction.h
@@ -180,7 +180,7 @@ struct prop *chrSpawnAtPad(struct chrdata *chr, s32 body, s32 head, s32 pad, u8 
 struct prop *chrSpawnAtChr(struct chrdata *basechr, s32 body, s32 head, u32 chrnum, u8 *ailist, u32 spawnflags);
 s16 chrGoToCover(struct chrdata *chr, u8 speed);
 void chrAddTargetToBdlist(struct chrdata *chr);
-s32 chrGetDistanceLostToTargetInLastSecond(struct chrdata *chr);
+f32 chrGetDistanceLostToTargetInLastSecond(struct chrdata *chr);
 bool chrIsTargetNearlyInSight(struct chrdata *chr, u32 distance);
 bool chrIsNearlyInTargetsSight(struct chrdata *chr, u32 distance);
 s16 *teamGetChrIds(s32 team_id);
